tutorials/rag_in_action/requirements.txt
openai

llama-index
llama-index-embeddings-huggingface
llama-index-llms-deepseek

langchain
langchain-openai
langchain-deepseek
langchain-community
langchain-huggingface
sentence-transformers

beautifulsoup4

python-dotenv

pymilvus
